gift问题修复

+++ Module/PlayWays/src/main/java-gift/com/module/playways/room/gift/GiftLocalApi.java
@@ -38,7 +38,7 @@ public class GiftLocalApi {
         getGiftDBDao().deleteAll();
     }
 
-    public void deleteById(long id) {
+    public void deleteById(int id) {
         getGiftDBDao().deleteByKey(id);
     }
 }
